added filtering events functionality

+++ Dto/Event/PartyEventDto.cs
@@ -5,7 +5,7 @@
         public string? Name { get; init; }
         public string? Description { get; init; }
         public string? Genre { get; init; }
-        public double? Price { get; init; }
+        public int? Price { get; init; }
         public string? ImageUrl { get; init; }
         public DateTime? EventTime { get; init; }
     }
