+++ Models/PartyEvent.cs
@@ -8,7 +8,7 @@ namespace night_life_sk.Models
         public string? Name { get; set; }
         public string? Description { get; set; }
         public string? Genre { get; set; }
-        public double? Price { get; set; }
+        public int? Price { get; set; }
         public string? ImageUrl { get; set; }
         public DateTime? EventTime { get; set; }
         public virtual PartyPlace? PartyPlace { get; set; }
